# apps/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.FileRequest:
    properties:
      content_type:
        type: string
      filename:
        type: string
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.PresignedURLRequest:
    properties:
      files:
        items:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.FileRequest'
        type: array
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.PresignedURLResponse:
    properties:
      uploads:
        items:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.UploadInfo'
        type: array
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.UploadInfo:
    properties:
      expires_at:
        type: string
      file_url:
        type: string
      id:
        type: string
      upload_url:
        type: string
    type: object
...
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_order_service.CreateOrderInput:
    properties:
      coupon_id:
        type: integer
      quantity:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for RevieU
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RevieU Core API
  version: "1.0"
paths:
...
  /media/{id}/analysis:
    post:
      description: Triggers analysis for a media upload
      parameters:
      - description: Upload ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Analyze media upload
      tags:
      - media
  /media/presigned-urls:
    post:
      consumes:
      - application/json
      description: Generates presigned URLs for uploading files directly to R2 storage
      parameters:
      - description: Files to upload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.PresignedURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_media_dto.PresignedURLResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create presigned URLs for media upload
      tags:
      - media
  /media/uploads:
    post:
      description: Creates a media upload and returns upload URLs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create media upload
      tags:
      - media
...
  /merchants/{id}/follow:
    delete:
      description: Unfollow a merchant
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unfollow merchant
      tags:
      - follow
    post:
      description: Follow a merchant
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Follow merchant
      tags:
      - follow
...
  /orders:
    get:
      description: Returns orders for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List orders
      tags:
      - order
    post:
      consumes:
      - application/json
      description: Creates a new order for the authenticated user
      parameters:
      - description: Create order request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_order_service.CreateOrderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create order
      tags:
      - order
  /orders/{id}:
    get:
      description: Returns an order by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get order detail
      tags:
      - order
  /orders/{id}/pay:
    post:
      description: Simulates payment success for an order and issues vouchers
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Pay order
      tags:
      - order
...
  /users/{id}/follow:
    delete:
      description: Unfollow a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unfollow user
      tags:
      - follow
    post:
      description: Follow a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Follow user
      tags:
      - follow
...
schemes:
- http
- https
swagger: "2.0"